# public/api-documentation.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Rempah API Documentation",
    "description": "Complete API documentation for the Rempah backend system including Customers, Orders, Order Items, Invoices, Receipts, and more.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@rempah.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api",
      "description": "Local Development Server"
    },
    {
      "url": "https://your-production-url.com/api",
      "description": "Production Server"
    }
  ],
...
  "paths": {
...
    "/orders/{order}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get order by ID",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "order",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Order retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {"$ref": "#/components/schemas/ApiResponse"},
                    {
                      "type": "object",
                      "properties": {
                        "data": {"$ref": "#/components/schemas/Order"}
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Orders"],
        "summary": "Update order",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "order",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "customer_id": {"type": "string"},
                  "order_date": {"type": "string", "format": "date"},
                  "remarks": {"type": "string"},
                  "discount": {"type": "number"},
                  "tax1_percentage": {"type": "number"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order updated successfully"
          }
        }
      }
    },
    "/orders/{id}": {
      "delete": {
        "tags": ["Orders"],
        "summary": "Delete order",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Order deleted successfully"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Sanctum Token"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "properties": {
          "status": {"type": "integer", "example": 200},
          "message": {"type": "string", "example": "Success"},
          "data": {"type": "object"}
        }
      },
...
      "Order": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "type": {"type": "string", "example": "SO"},
          "reference_no": {"type": "string"},
          "branch_id": {"type": "string"},
          "customer_id": {"type": "string"},
          "customer_code": {"type": "string"},
          "customer_name": {"type": "string"},
          "order_date": {"type": "string", "format": "date"},
          "status": {"type": "string", "example": "pending"},
          "gross_amount": {"type": "number"},
          "tax1": {"type": "number"},
          "tax1_percentage": {"type": "number"},
          "grand_amount": {"type": "number"},
          "net_amount": {"type": "number"},
          "discount": {"type": "number"},
          "remarks": {"type": "string"},
          "items": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/OrderItem"}
          },
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "unique_key": {"type": "string"},
          "reference_no": {"type": "string"},
          "order_id": {"type": "integer"},
          "item_count": {"type": "integer"},
          "product_id": {"type": "integer"},
          "product_no": {"type": "string"},
          "product_name": {"type": "string"},
          "description": {"type": "string"},
          "sku_code": {"type": "string"},
          "quantity": {"type": "number"},
          "unit_price": {"type": "number"},
          "discount": {"type": "number"},
          "amount": {"type": "number"},
          "is_free_good": {"type": "boolean"},
          "is_trade_return": {"type": "boolean"},
          "trade_return_is_good": {"type": "boolean"},
          "item_group": {"type": "string", "example": "Cash Sales"}
        }
      },
...
    }
  }
}